feat: Get coin working correctly
Re-worked the links between the Coin and CoinPoof Prefabs. That gave
me the correct access to the Coin.cs script and the Coin.OnClicked
method. The previous Coin GameObject has been replaced in the maze and
the Coins GameObject has been deleted. The Coin animations now work
when the Coin is clicked on in the game. Big thanks to Abigail the
forum mentor on the Udacity VR Nanodegree!